pyproject.toml
[project]
name = "nvidia-mim"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "numpy>=2.4.3",
    "nvidia-riva-client>=2.25.0",
    "openai>=2.29.0",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "requests>=2.32.5",
]
